+++ RELEASE_NOTES.md
@@ -5,6 +5,8 @@
 This is a compatibility release for sda-core 2.2.x.
 
 - `Changed` NodeType.getName to NodeType.getTypeName.
+- `Renamed` be.baur.sds.serialization.Parser to SchemaParser.
+- `Renamed` SchemaException to SDSParseException.
 
 ## Compatibility